+++ LolIn/src/main/java/com/iris/service/RepleServiceImpl.java
@@ -7,10 +7,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.iris.config.Config;
 import com.iris.dao.RepleDao;
 import com.iris.entities.Board;
 import com.iris.entities.Reple;
+import com.iris.libs.TrippleDes;
 import com.iris.utils.RepleQueryDsl;
+import com.iris.utils.SignatureUtil;
 import com.iris.vo.RepleVO;
 
 @Service
@@ -24,6 +27,8 @@ public class RepleServiceImpl implements RepleService {
 	RepleDao repleDao;
 	@Autowired
 	RepleQueryDsl repleQueryDsl;
+	
+	TrippleDes trippleDes;
 	
 	/**
 	 * 모든 댓글 조회
@@ -39,10 +44,21 @@ public class RepleServiceImpl implements RepleService {
 	 * 댓글 저장
 	 */
 	@Override
-	public String save(int boardId , String userName ,String content , String facebookId) {
+	public String save(int boardId , String userName ,String content , String facebookId ,String hash) {
+		
+		if(!SignatureUtil.compareHash(boardId+ userName + content + facebookId + Config.KEY.SECRET, hash)){
+			return null;
+		}
 		
 		Reple reple = new Reple();
 		Board board = new Board();
+		
+		try {
+			trippleDes = new TrippleDes();
+			facebookId = trippleDes.decrypt(facebookId);
+		} catch (Exception e1) {
+			e1.printStackTrace();
+		}
 		
 		board.setId(boardId);
 		reple.setAddBoards(board);
@@ -59,7 +75,11 @@ public class RepleServiceImpl implements RepleService {
 	 * 게시판의 댓글 조회
 	 */
 	@Override
-	public List<Map<String,Object>> findById(int boardId) {
+	public List<Map<String,Object>> findById(int boardId , String hash) {
+		
+		if(!SignatureUtil.compareHash(boardId + Config.KEY.SECRET, hash)){
+			return null;
+		}
 		
 		List<Reple> repleList = repleQueryDsl.findById(boardId);
 		RepleVO repleVO = new RepleVO();
@@ -70,7 +90,12 @@ public class RepleServiceImpl implements RepleService {
 	 * 댓글 삭제
 	 */
 	@Override
-	public String delete(int repleId) {
+	public String delete(int repleId , String hash) {
+		
+		if(!SignatureUtil.compareHash(repleId + Config.KEY.SECRET, hash)){
+			return null;
+		}
+		
 		repleDao.delete(repleId);
 		return DELETE;
 	}
